requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-analytics-data>=0.18.0
openpyxl>=3.1.2
pandas>=2.0.0
Jinja2>=3.1.2
requests>=2.31.0
